+++ Main/Features/MainMenuCustomizer/IMainMenuCustomizerService.cs
@@ -1,0 +1,6 @@
+namespace TAOM.Features.MainMenuCustomizer;
+
+public interface IMainMenuCustomizerService
+{
+    void CustomizeMenu();
+}